requirements.txt
flask==2.3.3
flask-cors==4.0.1
flask-limiter==3.5.0
pymongo==4.6.3
pyjwt==2.8.0
pandas==2.2.2
numpy==1.26.4
torch==2.3.1
open-clip-torch==2.24.0
ultralytics==8.2.48
transformers==4.41.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pytesseract==0.3.10
pillow==10.3.0
opencv-python==4.9.0
scikit-learn==1.4.2